+++ test/widgets/src/clide_markdown_test.dart
@@ -3,7 +3,6 @@
 library;
 
 import 'package:clide/widgets/widgets.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../../helpers/kernel_fixture.dart';